refactor all lines
