--- src/core/models/mod.rs
@@ -1,5 +1,0 @@
-pub mod auth;
-pub mod endpoints;
-pub mod timeline_v1;
-pub mod timeline_v2;
-pub mod tweets;
